requirements.txt
# Web自动化
playwright==1.51.0

# AI和机器学习
openai==1.12.0
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0

# 数据可视化
matplotlib==3.8.2
seaborn==0.13.2

# Binance API
python-binance==1.0.19

# 工具库
python-dotenv==0.19.0
requests==2.26.0
beautifulsoup4==4.12.3
schedule==1.2.1
aiohttp==3.9.1
typing-extensions==4.9.0

# 开发工具
pytest==8.0.0
black==24.1.1
flake8==7.0.0

# Web框架
flask==2.3.3
werkzeug==2.3.7 